Remove spice count. Implement card pool
Turns out tracking spice is illegal, so I've removed it from the game.
However, I've added a card pool - so now when you add cards to players,
it removes them from the pool, so it should be easier to figure out what
the unknown cards are based on the cards left in the pool.